refactor: Idle S3 bucket scanner
* clean dedup spinner
