+++ pkg/formatter/ebs_table.go
@@ -29,11 +29,6 @@ func PrintVolumesTable(volumes []models.VolumeInfo, scanTime time.Time, scanDura
 	// kubectl 스타일 tabwriter 설정
 	w := tabwriter.NewWriter(os.Stdout, 0, 8, 2, ' ', 0)
 
-	// Print scan timestamp first
-	fmt.Fprintf(w, "Scan time: %s (completed in %.2f seconds)\n",
-		scanTime.Format("2006-01-02 15:04:05"),
-		scanDuration.Seconds())
-
 	// Print header as requested
 	fmt.Fprintln(w, "NAME\tVOLUME ID\tTYPE\tSIZE\tSTATUS\tMONTHLY SAVINGS\tPRICING")
